Gegner können sterben, Gegner droppen ihr zeug armour = double

+++ src/item/Defense.java
@@ -5,7 +5,7 @@ import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.image.Image;
 
 public abstract class Defense extends Item{
-	private int armor;
+	private double armor;
 	
 	public Defense(String name, String description, int price, Image image, int x, int y, GraphicsContext graphicsContext, int armor) {
 		super(name, description, price, image, new Point2D(x, y), graphicsContext);
@@ -16,7 +16,7 @@ public abstract class Defense extends Item{
 		
 	}
 	
-	public int getArmor() {
+	public double getArmor() {
 		return armor;
 	}
